Move to begin/end transaction around voice initiation
This changes the responder API somewhat from count voices / make voices
to begin transaction / make voices / end transaction with the guarantee
that we will not begin another transaction until end is called. This allows
responders to share internal state over the voice creation while still
allowing the manager a chance to steal etc... although it does change
the public API of the responder.